huge Makefile fix, updated README

+++ cpp00/ex00/megaphone.cpp
@@ -8,7 +8,7 @@ int	main(int ac, char **av) {
 	else {
 		for (int i = 1; i < ac; i++) {
 			curr.assign(av[i]);
-			for (int j = 0; j < curr.length(); j++)
+			for (int j = 0; (size_t)j < curr.length(); j++)
 				std::cout << (char)toupper(curr[j]);
 		}
 	}
